requirements.txt
requests
feedparser
newspaper3k
readability-lxml
beautifulsoup4
textblob
pymongo
tqdm
python-dateutil
python-dotenv

# Integração com bancos de dados
psycopg2-binary
sqlalchemy
snowflake-connector-python
snowflake-sqlalchemy

# API de dados financeiros
yfinance
pandas
